Login page and student page
Added the main login page (index.html), the create new account page (links.html and create_account.php), and the main page for students (student.php and view_courses.php). To link the instructor and admin pages to the login, go to log_in.php and make sure the location of the header matches the name of the file for the account type.
e.g. 
case "admin":
				header("Location: admin.php");
				break;