Database V1 and view_courses patch
Fixed the sql table so that it gives 5 rows for each table and can be directly imported into the phpadmin database. It will delete any custom made data rows so save those if you need them. Also added a case in view_courses.php if a student has an A+ because I saw in the take table that a student can have an A+ in a class.