Database V2 and fixed GPA calculation
Added sections Section101 for courses COMP1010 and MATH1310 in the current year (2025) and semester (Spring). Also added ten new cs major students who are all taking the new cs section and the first half are taking the new calculus I section. Also added a new PhD student who isn't a TA. This was all done to properly test the functionality of Tasks 6 and 7. Finally, fixed a bug where a student could enroll in previous sections, and made it so that a Student that doesn't have a grade in a section won't have it count towards their GPA calculation.